Full Program
The code includes the algorithm to find the index of a desired number in a rotated, ordered array and the main.
In the main, I have it so it is one array and the algorithm is implemented for every single rotation possibility for that array. In each rotation, it is asked to look for the numbers 7 and 31. There is another array at the end that the algorithm also runs on to show it works.